Replace assert with debug_assert
I was under the assumption that asserts only panic in release builds. In
Rust this is not the case, but there's the debug_assert version. It only
panics in debug builds, not release builds.
This is the behavior I want because I want the application to gracefully
exit, instead of panic.

Signed-off-by: Daniel Schaefer <dhs@frame.work>