Implement dumping of bitmap from winux capsule
The capsule of type Windows UX capsule have a bitmap image embedded in
them. The UEFI implementation takes this image and displays it during
the firmware update.
Now we can extract it to look at it. THe next step would be to replace
it with a custom one.

Example invocation

```
> cargo build && cargo run -q -- --capsule winux.bin --dump dump.bmp
```

Signed-off-by: Daniel Schaefer <dhs@frame.work>