Add command to toggle PS2 emulation
The EC detects if the host can communicate with the touchpad via I2C and if not,
it'll act as an intermediary by emulating a PS2 mouse.
This is necessary for the Windows installer because Microsoft doesn't ship the Intel I2C driver.

This command is not necessary during normal use. In rare cases the EC might be stuck in emulation state
and this command can be used to set it properly. PS2 emulation doesn't support all the features of
the normal I2C protocol, for example multitouch gestures.

Signed-off-by: Daniel Schaefer <dhs@frame.work>