chromium_ec: Fix reading EC console
Now it can get all of the EC console buffer.
I see ~100 lines. Much more than the previous ~10 lines.

Follow also seems to properly work now.

A good way to force messages to appear on the console is to plug in or
unplug a charger, or to close and open the lid.

TEST=`framework_tool --console recent` shows the same output as `ectool console`
TEST=`framework_tool --console follow` keeps properly refreshing
endlessly.

Signed-off-by: Daniel Schaefer <dhs@frame.work>