Don't require PD config if platform isn't detected
Now even if the platform isn't detected, we can run the commands just
fine. For example both of these work (if I comment out the smbios entry):

```
framework_tool --versions
framework_tool --versions --pd-ports 1 2 --pd-addrs 66 64
```

The first one just fails accessing I2C passthrough for the PD ports.

This is the final change to allow just running the tool on any system in
as many cases as possible - even if the linux kernel or framework_tool
dose not recognize the system.

Signed-off-by: Daniel Schaefer <dhs@frame.work>