esrt: fix ESRT GUID for EFI device
The EFI GUID in FreeBSD is layouted as following

typedef struct efi_guid {
	uint32_t  Data1;
	uint16_t  Data2;
	uint16_t  Data3;
	uint8_t   Data4[8];
} efi_guid_t;

Which means the Data1, Data2, Data3 should follows the little endian
rule as the MSB should be in the right side.

Also, the table_len in the ioctl argument of EFI_GET_TABLE should be
follow the uuid structure. Fix it by reorder it.

Sponsored By:   FreeBSD Foundation
Sponsored By:   Framework Laptop Inc.